refactor: Bootstrap the project using standalone APIs
