+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { DOCUMENT } from '@angular/common';
+import { DOCUMENT, NgIf, NgClass } from '@angular/common';
 import {
   Component,
   ElementRef,
@@ -20,11 +20,17 @@ import { DomUtils, userAgent } from '@igo2/utils';
 import { delay, first } from 'rxjs';
 
 import { PwaService } from './services/pwa.service';
+import { PortalComponent } from './pages/portal/portal.component';
+import { MenuComponent } from './pages/menu/menu.component';
+import { HeaderComponent } from './pages/header/header.component';
+import { SpinnerComponent, SpinnerActivityDirective } from '@igo2/common';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss'],
+    standalone: true,
+    imports: [SpinnerComponent, SpinnerActivityDirective, NgIf, HeaderComponent, MenuComponent, PortalComponent, NgClass]
 })
 export class AppComponent implements OnInit {
   private themeClass = 'qcca-theme';
